requirements.txt
Flask
requests
Flask-SQLAlchemy
Flask-Login
Werkzeug
openai
python-docx
PyPDF2
pymysql
python-dotenv
azure-storage-blob